backend/pyproject.toml
[tool.poetry]
name = "fastapi-demo"
version = "0.1.0"
description = ""
authors = ["Deniz Seyidoglu <denizsydgl@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
gino = {extras = ["starlette"], version = "^1.0.1"}
alembic = "^1.7.4"
psycopg2 = "^2.9.1"
uvicorn = "^0.15.0"
gunicorn = "^20.1.0"
fastapi = "<0.70.0"
bcrypt = "^3.2.0"
PyJWT = "^2.3.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
requests = "^2.26.0"
black = "^21.10b0"
mypy = "^0.910"
pylint = "^2.11.1"
ipdb = "^0.13.9"
ipython = "^7.29.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."fastapi_demo.modules"]
"users" = "fastapi_demo.features.users.app"
"tasks" = "fastapi_demo.features.tasks.app"
"songs" = "fastapi_demo.features.songs.app"
"comments" = "fastapi_demo.features.comments.app"
